refactor: Move T4C repository models to repository module
