+++ backend/capellacollab/settings/modelsources/t4c/repositories/models.py
@@ -3,6 +3,7 @@
 
 from __future__ import annotations
 
+import enum
 import typing as t
 
 from pydantic import BaseModel
@@ -10,6 +11,8 @@ from sqlalchemy import Column, ForeignKey, Integer, String, UniqueConstraint
 from sqlalchemy.orm import relationship
 
 from capellacollab.core.database import Base
+from capellacollab.core.models import ResponseModel
+from capellacollab.settings.modelsources.t4c.models import T4CInstance
 
 
 class DatabaseT4CRepository(Base):
@@ -25,3 +28,34 @@ class DatabaseT4CRepository(Base):
     instance = relationship(
         "DatabaseT4CInstance", back_populates="repositories"
     )
+
+
+class CreateT4CRepository(BaseModel):
+    name: str
+
+
+class T4CRepositoryStatus(str, enum.Enum):
+    ONLINE = "ONLINE"
+    OFFLINE = "OFFLINE"
+    INSTANCE_UNREACHABLE = "INSTANCE_UNREACHABLE"
+    NOT_FOUND = "NOT_FOUND"
+
+
+class T4CRepositories(ResponseModel):
+    payload: t.List[T4CRepository]
+
+
+class T4CInstanceWithRepositories(T4CInstance):
+    repositories: list[T4CRepository]
+
+    class Config:
+        orm_mode = True
+
+
+class T4CRepository(CreateT4CRepository):
+    id: int
+    instance: T4CInstance
+    status: t.Optional[T4CRepositoryStatus]
+
+    class Config:
+        orm_mode = True
